requirements.txt
click
pandas
streamlit
protobuf
python-dotenv
rapidfuzz
google-generativeai